dj_nnapi/requirements.txt
albucore==0.0.24
albumentations==2.0.8
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.9.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Django==5.2.6
django-cors-headers==4.8.0
django-debug-toolbar==6.0.0
django-filter==25.1
django-prometheus==2.4.1
django-use-email-as-username==1.4.0
django_dramatiq==0.14.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dramatiq==1.18.0
dramatiq-dashboard==0.4.0
drf-yasg==1.21.10
efficientnet-pytorch==0.7.1
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.9.0
gunicorn==23.0.0
huggingface-hub==0.34.4
idna==3.10
imagecodecs==2025.8.2
imageio==2.37.0
imgaug==0.4.0
inflection==0.5.1
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
kombu==5.5.4
lap==0.5.12
lazy_loader==0.4
MarkupSafe==2.0.1
matplotlib==3.10.6
mpmath==1.3.0
munch==4.0.0
networkx==3.5
numcodecs==0.16.2
numpy==2.2.6
nvidia-ml-py==13.580.82
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pika==1.3.2
pillow==11.3.0
polars==1.33.1
pretrainedmodels==0.7.4
prometheus_client==0.22.1
prompt_toolkit==3.0.52
psutil==7.0.0
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic_core==2.33.2
pydicom==3.0.1
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==4.5.4
requests==2.32.5
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.16.2
segmentation-models-pytorch==0.3.4
shapely==2.1.1
simsimd==6.5.3
six==1.17.0
sqlparse==0.5.3
stringzilla==4.0.14
sympy==1.14.0
tifffile==2025.9.9
timm==0.9.7
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.203
ultralytics-thop==2.0.17
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
xgboost==3.0.5
